% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo_add.R
\name{cargo_add}
\alias{cargo_add}
\title{cargo_add}
\usage{
cargo_add(project, package, additional = NULL)
}
\arguments{
\item{project}{chr the project names with dirs}

\item{package}{chr the package need to used}

\item{additional}{chr additional parameters in build}
}
\value{
web
}
\description{
the "cargo add " command
}
