% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo_doc.R
\name{cargo_doc}
\alias{cargo_doc}
\title{cargo_doc}
\usage{
cargo_doc(project, additional = NULL)
}
\arguments{
\item{project}{chr the project names with dirs}

\item{additional}{chr additional parameters in build}
}
\value{
web
}
\description{
check the doc of used rust package
}
