% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo_new.R
\name{cargo_new}
\alias{cargo_new}
\title{cargo_new}
\usage{
cargo_new(project, lib = FALSE, root = ".")
}
\arguments{
\item{project}{chr the project names with dirs}

\item{lib}{chr the "--lib" parameter}

\item{root}{chr the root dir of project,default as "."}
}
\value{
project
}
\description{
the "cargo new" command
}
\examples{
hello_world <- cargo_new("dev/hello_world")
}
