% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo_run.R
\name{cargo_run}
\alias{cargo_run}
\title{cargo_run}
\usage{
cargo_run(project, additional = NULL)
}
\arguments{
\item{project}{chr the project names with dirs}

\item{additional}{chr additional parameters in build}
}
\value{
project
}
\description{
the "cargo run "command
}
\examples{
cargo_run("dev/hello_world")
}
