% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo_test.R
\name{cargo_test}
\alias{cargo_test}
\title{cargo_test}
\usage{
cargo_test(project, additional = NULL)
}
\arguments{
\item{project}{chr the project names with dirs}

\item{additional}{chr additional parameters in build}
}
\value{
project
}
\description{
the "cargo test "command
}
\examples{
cargo_test("dev/hello_world")
}
