% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cargo.R
\name{install_cargo}
\alias{install_cargo}
\title{install_cargo}
\usage{
install_cargo(project, additional = NULL)
}
\arguments{
\item{project}{chr the package names}

\item{additional}{chr additional parameters in build}
}
\value{
null
}
\description{
the "cargo install "command
}
