% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst_compile.R
\name{typst_compile}
\alias{typst_compile}
\title{typst_compile}
\usage{
typst_compile(ptyfiles, pdfdir = ".", watch = FALSE, font_path = "")
}
\arguments{
\item{ptyfiles}{chr the dir of ptyfiles}

\item{pdfdir}{chr the dir of output pdf file}

\item{watch}{logit Watches source files and recompiles on changes when TRUE}

\item{font_path}{chr the diy font path}
}
\value{
null
}
\description{
compile typst in cargo
}
