% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_cargo.R
\name{uninstall_cargo}
\alias{uninstall_cargo}
\title{uninstall_cargo}
\usage{
uninstall_cargo(project, additional = NULL)
}
\arguments{
\item{project}{chr the package names}

\item{additional}{chr additional parameters in build}
}
\value{
null
}
\description{
the "cargo uninstall "command
}
